Renamed and reworked AttachToCouple to CoupledAttachment.
