+++ CCL.Types/Components/CoupledAttachmentTag.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace CCL.Types.Components
+{
+    public class CoupledAttachmentTag : MonoBehaviour
+    {
+        public string Tag = string.Empty;
+    }
+}
